Prepare for converting to Kokoro's new Docker pool.
They're deprecating the current non-docker one in July and keep sending me CL's to migrate off it.

The Kokoro-setup script has been modified to check which pool we're running on and run the unique configuration for each.

Also included a basic Dockerfile that I used for testing locally. It'll be great to turn this into something more fully featured with pre-cached downloads and dependencies.

BUG: b/274946478
Change-Id: I197f547163c76fde6f36a7e3b8d018c534acd1ad