Dial and Gauge Support
Create a new DesignCompose plugin that allows configuration of a Figma node to be used as a dial, gauge, or progress bar.

- Arc customizations allow an ellipse to be configured with a start and end angle. The ellipse can optionally have a stroke with rounded corners, which is saved only in the plugin data since it is not exposed from the REST API.
- Rotation customizations allow any node to be configured with a start and end rotation.
- Progress bar customizations allow a frame or rectangle vector to be configured with a start and end position, and horizontal or vertical.

Each of these customizations has a slider to let the designer view the dial/gauge in different positions. This data is written into the shared plugin data of the node as "meter-data".

Bug: 272365059

Support rendering of dials and gauges in DesignCompose

Rust:
Parse the meter data from the new plugin. To render ellipses as a dial, the rendering code needs to draw the ellipsis arcs manually instead of from the list of paths provided by Figma, so we parse additional arc data from ELLIPSE type nodes and save them into a new "Arc" ViewShape. Similarly, rectangular vectors for progress bars are also saved into a new "VectorRect" ViewShape.

Kotlin:
Add a new "Meter" customization to configure a node as a meter. This is just a Float value that should be set to be between 0 and 100. When rendering the node in DesignFrame.frameRender(), use the meter data to render the meter appropriately.

- Progress bar: Adjust the width or height of the frame according to the meter value.
- Rotation: Rotate the node about its center according to the meter value. Rotations are a bit tricky because the transform matrix also contains translation values from the node's parent, so we undo the translation, rotate the node about its center, then redo the translation.
- Arc: Instead of using the list of paths provided by Figma, we need to generate our own path for arc meters. Render the outline of the arc if it has an inner radius; otherwise render just the stroke of the arc. For arcs with rounded corners, construct cubic bezier curves.

Change-Id: Iff60d70dbc33372861bb46efa9b2e522f7802f0f