Fix a crash in the list widget when formatting is set to "Hug contents".
When overflow scrolling is enabled and then the format is changed to "Hug contents", the scrolling boolean value was getting set to false after saving the plugin data, causing the previous true value to get saved and used in a running application. Fix this by setting scrolling value to false first.

This commit also removes some unused code.

This commit also changes some plugin html code to set the 'text' field instead of 'innerHTML' field in an attempt to fix a security issue.

Bug: 278312107

Change-Id: Ia168fba60ad71264ce77d2a85e4eaa049096c257