Add support for injecting fake key events to trigger Figma interactions
Figma's interaction tools give us key input as a way to trigger an action. This commit adds a function to inject one of these key events, which consists of a key and a list of optional metakeys (shift, alt, control, meta). There are two ways to use this key injection function.

1. Call it directly with the DesignInjectKey() function.
2. Create a new function annotated with the @DesignKeyAction() annotation. This generates a function that calls the DesignInjectKey() function with the parameters specified in the annotation.

An example of this has been added to the Interaction validation document.

The way this works is that when a node with key reactions is rendered, we register a function with the node as a listener for the key events it is interested in. When the key injection happens, listeners are notified and their interactions are dispatched.

Bug: 277762504