Migrate the buildSrc code to it's own Gradle subproject. (#9)
BuildSrc is a special thing in Gradle, which is supposed to
hold build configuration that's common to your subprojects.
It's essentially it's own build, but always had some odd edge cases

While working on the rust plugin I found that plugins must be in
their own project and included as an includeBuild. Since the plugin
needed some of the configuration in buildSrc, the change needs to be
made. Splitting this out ahead of the rust plugin PR