Remove Standalone Validation app
PR #13 makes the Tutorial app into it's own project. I used a new method that makes things a bit cleaner and avoids some (but not all) of the symlinking.

The standalone Validation app basically existed to make sure the published repo worked, but the new Tutorial app handles that better. So I'm removing the standalone version of the Validation app.

Meanwhile, MediaCompose  (which is now the only app in the AAOS Unbundled project) works the same way now, basically. So I can remove it's "Standalone" version.