Replace the existing Rust Android Gradle Plugin with our own (#11)
NOTE: This requires a rust upgrade to 1.68. Run `rustup update`

The new plugin includes a few new things vs the previous one:
- Stable interface with Android Gradle Plugin
The plugin is properly integrated with AGP's build lifecycle and
uses Gradle's and AGP's current APIs for creating task dependencies
- Configured for Gradle Build Caching
Gradle is aware of whether the sources for the build have changed and
whether it needs to run the task and trigger Cargo
- Handles Debug and Release builds properly
Creates individual tasks for compiling the Rust code with and without
the --release flag, using the right build profile for the right
Gradle artifact
- Stable with AGP 8+
The previous plugin, starting with AGP 7.4, would sometimes
build an AAR that was missing the JNI libraries.
- Removes the workarounds that were required to compile Rust against
the Android NDK
See https://blog.rust-lang.org/2023/01/09/android-ndk-update-r25.html