Security code scan security warnings
When saving the overflow node preview image data, just save the binary data without the html src header. Then explicitly set the image to the concantenation of a manually written header and the saved binary image data.

This commit also fixes a small bug found with enabling the "Limit elements" section where the content did not display when switching the section on.

Fixes #59