Support progress bars indicator mode
Support a new progress bar mode called "indicator mode" where an element just moves along the bar without changing size.

This commit also changes the way you setup a progress bar in Figma. Instead of setting the start/end values in the plugin, these values are now automatically calculated from the parent frame. A parent frame is required for a node to be used as a progress bar, and the plugin UI prohibits the progress bar switch unless a parent frame exists. A tooltip has been added to explain this.

The plugin data written into the Figma nodes has been changed accordingly, as well as the code to render it in FrameRender.

A number of typescript warnings have been fixed as well.

Fix #58