Removing build results tracking from the CargoBuild Task  (#29)
Stacked on #24

Attempting to have it be cacheable and attempting to determine if we can
skip running Cargo was pretty naive. Trying to get it right will be
super error prone.

Cargo is already pretty quick at returning if it doesn't need to build.
Gradle adds to that time, but once we have configuration caching (#26)
enabled, the tasks will run simultaneously and won't really add to the
build time.

Fixes Investigate possible caching issue with Rust serialization #27