Enable presubmits to run on pull requests targeting working branches (#33)
This will make it easier to stack pull requests. This is where you've
uploaded a pull request, then create a branch from that pull request,
make changes, and upload that pull request, with the merge base being
the pull request that was branched from.

With this change, the presubmits will properly run on the stacked PR.