Update tutorial file again with a better name (#57)
I realized in the design switcher it uses the name of the file, which in
this case was "#welcome-is-stage" or something like that. I rename the
branch that has #stage set to the welcome page to "Tutorial Welcome" so
that it looks better when you switch from the pre-loaded design to the
new copy.