Use a bundler for the plugin to separate images from code (#102)
Use webpack to keep images separate from code so that it is easier to do
code reviews and keep images separate. Webpack will inline the images
into the output code.

This also updates the README file since it was outdated and makes it
more consistent with the widget README.

Fixes #86