Propagate better errors from figma_import JNI (#48)
- I saw the JNI crate using the "thiserror" to help with error handling,
so I added it here to make it easier to display errors that we aren't
specifically handling

- Most of the errors from Rust should be handled by doing more than just
printing an error, so I'm setting up a mapping between Rust errors and
Exceptions that'll be thrown. I think that interpreting the errors into
simple exceptions should be handled on the Rust side, partly to reduce
the amount of data that needs to be copied over.

- I cleaned up the error display for Feedback a bit, making it more
concise.

- I've also updated the rust-JNI library to 0.21, which improves safety
especially related to reference lifetimes.