Adding Emulator testing to CI (#105)
Adding Emulator tests to CI, which will essentially start up an
[ATD](https://developer.android.com/studio/test/gradle-managed-devices#gmd-atd)
and run `./gradlew connectedCheck` on it.

As part of getting it running I needed to fix up some tests and update
the Figma docs for them.

I've also refactored the start-up and cache handling code for the GitHub
actions a bit, to make sure that we have a consistent environment with
the different jobs and make the caching more consistent for Android Rust
builds.

Also, to save some processor time and for a bit of simplicity, I'm
removing the macos runs from the presubmits. It should be fine to only
test them nightly.