Update Dials & Gauges plugin UI
Update the UI per latest designs. This adds images, tooltips, and incremental progress tick marks.

This commit slightly modifies how progress bar markers work. Progress bar marker positions are now calculated by simply moving them along the x-axis, with the range calculated using the parent frame's width and the marker's width. The marker now moves such that at the endpoints, the middle of the marker aligns with the edge of the parent.

This simplification relies on a change in how the parent frame of the marker is setup. Whereas previously the marker assumed that it had an axis aligned parent frame, it now assumes that the parent has the same skew and rotation transform as the marker. This means that in order to make a skewed or rotated progress bar, a designer would first create the marker and its parent as an axis aligned rectangle, then apply the skew/rotation to the parent, which affects the marker as well.

Fixes #116