Automate creation of release artifacts (#113)
This sets up a new workflow action that triggers when a new tag is
created, building the Maven repo and the Figma resources and uploading
them to the GitHub release for that tag. The Gradle build will have the
version set to the tag's name, so if a tag named "v0.20.0-rc03", the
DesignCompose libraries will be version "0.20.0-rc03". (No more making
commits to set the version to rc02 or rc03 or whatever).