Support masking layers from Figma (#85)
On the Rust side, parse the "is_mask" parameter from Figma from frames
and vectors that have this property, and export this into the
corresponding ViewShape that we write into a serialized Figma file.

In Kotlin, reorder children when masks are involved and keep track of
some new data to help render masks properly.

When a node has a child that is a mask, we give the node a MaskViewType
of MaskParent. This causes two changes when we render the node. First,
we save the size of the parent node's draw context so that when it's
child mask renders, the mask can make sure it masks the entire parent.
Second, we render the parent and all its children under a
withSaveLayer() call so that they are rendered in a separate layer to
ensure that the alpha channel of the mask is applied properly. When
rendering the children of this parent node, we first iterate through the
children and reorder them so that masks are rendered after the siblings
that they affect.

When a node is a mask, we give the node a MaskViewType of MaskNode. This
also causes two changes when we render the node. First, we get the saved
size of the parent node and use it to create a rectangle that matches
the parent's size. Second, we render the mask under a withSaveLayer()
call with blendmode set to DSTIN. This causes the mask to be rendered
with alpha blending only where the nodes that it masks were already
rendered.

Fixes https://github.com/google/automotive-design-compose/issues/63