Fix a bug where clicking "load" on the design switcher without setting an ID will soft-lock the switcher. (#143)
Check for empty string before executing a load in the design switcher.
This also prevents the design switcher overlay when clicking "load" with
an empty ID. Check for empty string before switching documents in the
design switcher as an extra safety check.

This commit also does a small amount of cleanup in the design switcher
code so that all the composable functions use the same LiveUpdateMode
setting.

Fixes #56