Append release versions to the generated artifacts (#145)
This will make it easier to tell the versions apart, and help with the
naming for uploading them to GCS (as part of #101)

Also moves the zipping of the plugin and widget from before the artifact
is uploaded to the job result to before uploading them to the release.
The upload-artifact job that uploads to the job result already zips up
the files, so it was unnecessary there.