Minor fix to make the ATD test plugin apply more consistently (#152)
The script for it would apply the plugin that adds the ATD tests, then
try to add some extra configuration to the tasks that were created. The
tasks are created by the Android Gradle Plugin, so to add that
configuration they have to be looked up manually. In some cases this
plugin was being applied to the project before the Android Gradle Plugin
was being applied, so the tasks were trying to be configured before they
were created. This fixes it to apply the configuration inside the
`withType()` function, which waits to apply the configuration until the
plugin with that type is applied to the project.