Add the DesignCompose plugin (#166)
Move the SetFigmaTasks to the plugin and enable it for each app.

This enables the Tutorial app to be full split from the main build and
for it to use the previous release of Android Studio, Electric Eel.

Tested by opening the Tutorial project in Electric Eel, building and
launching the app and setting the token.

I've not yet applied it to MediaCompose, though I may be able to do that
soon in a subsequent PR.

Closes #54
Closes #92