Retrieve component sets for any component instances found
While fetching and parsing nodes, the Figma to serialized doc process only stored full component set data when explicitly requested to do so via @DesignComponent annotations or the --nodes parameter from the fetch tool. This commit changes this so that if any component instances are encountered while parsing through nodes, retrieve full component set data and store it in the serialized file. This means that requesting a single node #Stage, for example, that has an instance of the component set #cluster/prnd in it, will result in the retrieval of the full component set for #cluster/prnd with all its variants.

Fixes #179