Fix a bug with rendering arcs that have an inner radius < 1.0 (#172)
When an arc with an arc meter customization has an inner radius < 1.0,
we manually draw a path around the perimeter of the arc using arcs,
bezier curves, and lines. The math to calculate all of these paths
assumed that the sweep angle (angle going from start to end) was always
positive, so when this was not the case the arc rendered incorrectly.
This commit fixes this issue by swapping the start and end angles when
the sweep angle is negative.

Fixes #147