Fix variant selection not propagating to children properly (#156)
When we use the @DesignVariant annotation to declare a variant
customization and then change the variant by passing in an enum
argument, the code takes a path in DesignFrame() where it resolves the
variant parameters and finds the correct node to actually compose. Then
it calls a customComposable() composable function that composes the
variant we want. There was a problem where the customComposable function
did not propagate all customizations down to its children properly:
specifically, it did not propagate the customComposable() function
itself. This has been fixed in the CustomizationContext.mergeFrom()
function.

The Variant Properties test has been modified to test for nested
variants. While testing this another issue came up where variants can
sometimes render two states at the same time: the one saved in Figma,
and the one set dynamically in code. This has been fixed by only calling
frameRender() if we are not calling the customComposable() function to
render a different variant.

Fixes #146