Update Dependabot config to try to fix Gradle updates (#293)
The Gradle dependencies haven't been updated by Dependabot in a while
and it seems related to some of the `group` and `ignore` options that I
tried to enable. I've filled
https://github.com/dependabot/dependabot-core/issues/7804 about iit but
haven't seen a response yet.

In the mean time I've disabled those, so I'll need to manually tell
Dependabot to ignore certain dependency updates that need to be done
manually or in concert with others. I will try to balance the amount of
PR spam by setting the schedule to "daily" but limit it to only 3 PRs
open.

I've also set the other package ecosystems to only group together
updates that update the patch number (as opposed to major or minor) of
the dependency, so that additional scrutiny can be applied to other
changes. They're also now spread out over the week, so they don't all
open a bunch of PRs every Monday.

(This change has come out of a lot of experimentation on my fork of the
Repo, so it should be good to go as is)