#315 Include original relative_transform in serialized output. (#322)
This is to address the issue I saw and filed as #315.

The primary change is to include the original unaltered
`relative_transform` in the serialized figma output. Since this addition
is merely additive, that change should not break anything.

~~**However** I'm also making a correction to the `transform`
calculation which I believe to be correct, but which could break
downstream clients. Maybe we can talk about how to test this.~~

The addition of the original `relative_transform` in the serialzied
document has allowed me to decrease the complexity of of our renderer
(slighly) and also fixes the original issue. I'll attach a screenshot of
our (non-flattened) telltales below.

![Screenshot from 2023-08-24
23-37-05](https://github.com/google/automotive-design-compose/assets/2950232/54c2fbe4-f8f8-4d96-9c93-ac2d43ced24f)

---------

Co-authored-by: John O'Neil <johnoneil@google.com>