Add support for testing live fetches from Figma
This a new field to DesignSettings which records when and how
a DesignDoc was last fetched. This allows for instrumented tests to
confirm that the file be rendered was fetched by Figma.

This also adds a Semantics tag that is applied to the DesignDoc's root DesignView.
This allows a tester to tell whether a DesignDoc is being rendered

A test fixture was added which exposes a method for enabling
live update within a test via a key provided via test arguments. The key
is not embedded in the app this way. At this time AGP doesn't support
compiling Kotlin test fixtures (long story) so it's written in Java.