Add support for testing live fetches from Figma
This involves adding a test fixture which exposes a method for enabling
live update within a test via a key provided via test arguments. The key
is not embedded in the app this way.

This also adds a new field to DesignSettings which records when and how
a DesignDoc was last fetched. This allows for instrumented tests to
confirm that the file be rendered was fetched by Figma.