Enhance DesignCompose testability (#352)
This adds a new field to DesignSettings which records when and how
a DesignDoc was last fetched. This allows for instrumented tests to
confirm that the file be rendered was fetched by Figma.

This also adds a Semantics tag that is applied to the DesignDoc's root
DesignView.
This allows a tester to tell whether a DesignDoc is being rendered

A test fixture was added which exposes a method for enabling
live update within a test via a key provided via test arguments. The key
is not embedded in the app this way. At this time AGP doesn't support
compiling Kotlin test fixtures (long story) so it's written in Java.

I ended up cleaning up DesignCompose's dependencies as well, there
should be nothing affected by this.

TO TEST:

We do not have a way to run Instrumentation tests or Figma fetches via
GitHub actions. The tests added here must be performed manually.

To watch and/or debug the tests: Launch an emulator and run `./gradlew
connectedTest`

To run the tests automatically using automated test devices: run
`./gradlew tabletAtdApi30DebugAndroidTest`
Edit: After #363 merges that will be `./gradlew gmdTestQuick`

To run the full test suite: run `./dev-scripts/test-all.sh`