Initial Roborazzi support (#359)
This adds a basic Roborazzi test to HelloWorld and includes
documentation on it. Subsequent PRs will add roborazzi tests for
Validation and the DesignSwitcher.

Please review the implementation and let me know if there's any
questions. Once approved I'll update the Validation draft with the
changes and finish the DesignSwitcher Roborazzi tests.

See the Readme and new docs for instructions.