Add GitHub workflows for creating screenshot diff comments (#389)
This adds workflows that implement automatic comments that show the
difference between the result of the PR and it's base branch. These are
based on the example here:
https://github.com/takahirom/roborazzi-compare-on-github-comment-sample

These workflows won't function until they're comitted to main, so you
won't see it working on this PR. See
https://github.com/timothyfroehlich/automotive-design-compose/pull/102
for an example of the workflows in action.