Optimize layout by reducing the number of calls to compute layout
Add a beginLayout() and endLayout() function to the LayoutManager that keeps track of layout IDs in progress. When a DesignFrame subscribes to layout, call beginLayout() which adds its layout ID to a set of IDs. After the frame's children composables have been called, call endLayout() which removes the layout ID from the set. When the set is empty, the frame and all its children have completed, so compute layout. This reduces layout computation in scenarios such as a LazyGridView with lots of children. This now computes layout for every child of the LazyGridView instead of every ancestor of every child.