Begin allowing Dependabot to update indirect dependencies of rust libs (#417)
PR #416 is failing because the new version of ureq won't build. On a
hunch I ran `cargo update` and that solved the problem (and some
packages that looked like they were related were indeed updated).

This change will include that update to the dependencies and configure
Dependabot to allow updates to the dependencies of our explicitly
required dependencies.