Add default ktfmt IntelliJ settings (#444)
These settings will enable the ktfmt plugin and set it to use the
correct format.

Note: This will override your default ktfmt plugin settings, however,
the only settings offered are off/on and which style to use. The plugin
should be on and it should be using the KotlinLang style.

(Also caught some whitespace formatting in the .gitignore).