Closes #401: Support FILL sizing mode for autolayout frames
When parsing a node from Figma with layoutGrow set to 1, also set flex_basis to 0. This lets the node start out small and expand to fit its parent.

This commit also adds unit tests for FILL frames, and two scripts for testsing: one to fetch and save the test Figma file, and one to run the test.