Change rendering tests to detect semantic on the root DesignDoc
Right now the RustLayout code is ignoring Modifiers (#458) which means the current way for tests to check that a doc was rendered won't work. This change adds a semantic to the generated DesignDoc interface on an empty Box so that we can always find the correct node.

This change also changes what we're detecting. Instead of a DocID (which won't always stay the same), we check the class name of the generated DesignDoc.