Closes #486: Reduce JNI add-view calls
When subscribing to layout, add the node to a list of layout nodes instead of adding them directly via a JNI call. When we detect we are done with layout and need to compute it, make one JNI call to add all the views and compute layout on the root node.

The function to compute layout now needs the layout id of the node to compute, instead of computing layout on all root level nodes. This helps performance by not computing layout on nodes that we know have not changed.