Change layout subscriptions to add ViewStyle instead of View
Subscribing to layout with a View caused unnecessary serialization of data. In particular, serializing a View also serialized all of a View's children, meaning we were serializing O(n^2) data that wasn't even used. Since the Rust layout only uses the style within View, and the name for debugging purposes, it now takes those as parameters.

When subscribing to layout, use the ViewStyle created from merging with overrideStyle instead of the style that came with the View originally. This fixes the style used for variants and replacement nodes and means we no longer need the "baseView" mechanism of overriding styles.

Add another "rootLayoutId" parameter to ReplacementContext for ContentReplacement customizations. Replacement customizations should use this in the "parentLayout" parameter for the replacement content.