Closes #464: Improve grid view load and scroll performance
When the lazy grid view first loads and when scrolling through it quickly, some of its children don't have their layout computed yet, resulting in a child that takes up no space, which makes the lazy grid view load more children. Sometimes at load time, all the children are loaded because of this.

This commit fixes this issue by having the list preview widget write size values for all nodes that can go in it. We parse this data out of the shared plugin data and then ultimately use it as the default/initial size of a node if its layout has not yet been computed.