Enable unit tests and Roborazzi screenshot tests in feature/rustlayout (#500)
Create tasks to build the Rust libraries for the developer's host,
configure the test tasks to use it

Three PR's in one, but it's simplest to have them together to have tests
working.

The first is the CargoPlugin changes that splits the CargoBuildTask into
Android and Host subclasses and registers tasks to build the libraries
for the host.

The second is the changes to the build.gradle.kts files which add
dependencies on the host build tasks and adds the libraries to the
`java.library.path` for the tests. This is very hacky and brittle and
I've mostly figured out how to do it properly, but it was taking too
long and I want to get screenshot tests working. I have #502 to do this
properly.

The last is getting the screenshot tests working, both updating the
serialized files and setting the tests to expect failure, so that the
tasks that generate and compare screenshots still succeed.

There won't be any screenshot diff comments in this PR, because the
comment workflow compares the screenshots generated from this PR against
those generated by the base branch, feature/rustlayout, which isn't yet
generating screenshots! But once this is merged into to
feature/rustlayout, the tracking PR for merging it into to main, #461,
will update with the screenshot comparison vs main. I put up #503 to
make sure the screenshot diffs were working too.

Fixes #484

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>