Remove action that sets auto-merge for Dependabot (#529)
Apparently this is what has been causing the Dependabot PRs to fail to
launch the CI workflows in a merge queue:
https://github.com/orgs/community/discussions/70310

A week ago I'd disabled the merge queue to try to resolve the issues
with Dependabot PRs not merging, but because we have a requirement for
PRs to be up to date with main, we still had to rebase and merge the PRs
one-by-one.

This change should resolve everything. I'll re-enable the merge queue
and manually click "merge" when I approve dependabot PRs. The'll go into
the queue and actually launch their actions, so all of them can merge
without me manually clicking each of them.