squoosh: implement nested composables with functional hit testing.
This isn't perfect -- transforms aren't considered for hit testing, but
are used for rendering. So rotated notes don't work as you might expect.
This is fixable by decomposing the transform and applying it to child
Composables when laying them out, but I chose not to implement that yet
(largely because none of the Android Matrix types have a decomposition
method implemented, and it's a bit detailed).