Closes #486, #487, #488, #490: Performance improvements (#491)
When subscribing to layout, add the node to a list of layout nodes
instead of adding them directly via a JNI call. When we detect we are
done with layout and need to compute it, make one JNI call to add all
the views and compute layout on the root node.

The function to compute layout now needs the layout id of the node to
compute, instead of computing layout on all root level nodes. This helps
performance by not computing layout on nodes that we know have not
changed.

When computing layout in Rust, send back not only the layout IDs that
changed, but all the layouts as well. The layout manager in Android now
caches these layouts, so retrieving them no longer requires a JNI call.

Change layout subscriptions to add ViewStyle instead of View
Subscribing to layout with a View caused unnecessary serialization of
data. In particular, serializing a View also serialized all of a View's
children, meaning we were serializing O(n^2) data that wasn't even used.
Since the Rust layout only uses the style within View, and the name for
debugging purposes, it now takes those as parameters.

When subscribing to layout, use the ViewStyle created from merging with
overrideStyle instead of the style that came with the View originally.
This fixes the style used for variants and replacement nodes and means
we no longer need the "baseView" mechanism of overriding styles.

Add another "rootLayoutId" parameter to ReplacementContext for
ContentReplacement customizations. Replacement customizations should use
this in the "parentLayout" parameter for the replacement content.

After computing layout for a specific layout_id, calculate layouts that
changed for only the specified layout_id and its children instead of all
root nodes.

Reduce layout computation when text nodes are measured. Simplify and
improve text measurement and layout by measuring text before subscribing
to layout, and then sending the text size when subscribing. This is an
improvement in a few ways:
- Every text node visible no longer triggers a layout recomputation
since they now subscribe in the same manner as frames where layout
computation is done after all nodes have subscribed
- Text is laid out at the same time as frames, avoiding the problem of
text flickering

Don't recompute layout when removing nodes that are ancestors of a list
widget. Nodes are frequently removed from lazy grid views as they scroll
and no layout needs to be recomputed since the lazy grid view handles
layout itself.

Add another optimization when removing nodes by only computing layout
for the root node of the node removed.