Closes #464: Improve grid view load and scroll performance (#524)
When the lazy grid view first loads and when scrolling through it
quickly, some of its children don't have their layout computed yet,
resulting in a child that takes up no space, which makes the lazy grid
view load more children. Sometimes at load time, all the children are
loaded because of this.

This commit fixes this issue by having the list preview widget write
size values for all nodes that can go in it. We parse this data out of
the shared plugin data and then ultimately use it as the default/initial
size of a node if its layout has not yet been computed.

Clean up some code relating to list widget
- Change list content customization so that the replaced
https://github.com/composable content function includes parent layout
info. This should be passed to the replacement content so that when it
is composed, DesignFrame knows it is the child of a widget and can
handle it differently.
- Remove unnecessary Box() under DesignFrameLayout()
- Clean up some unnecessary logging

Changed layout to use constraints. Our custom layout function now gets
the layout for a node's children and uses each child's width and height
to construct a fixed size constraint for it. This lets us remove the
special case Box() for children of a Grid, where there were some layout
oddities without this change.

Update serialized files and screenshots