Towards #464: Revert grid node size changes (#549)
The grid node sizes output from the list widget preview are no longer
needed. Previously, the children of lazy grids were a Box() with a size
retrieved from layout. The layout retrieval was being called in the
DesignFrame() composable before layout was complete, so layout could be
null. Now the children of lazy grids use our custom DesignFrameLayout,
where layout is retrieved in the layout phase after all the DesignFrame
composables have been run, so layout is not null.