Fix detection of app module names in Android Studio
I'm not 100% sure what's causing the issue, but either way, Android Studio seems to have issues distinguishing the names of the different `app` directories. Changing the directories names to something unique (such as `helloworld/app` to `helloworld/helloworld-app` seems to take care of the issue and Android Studio will name the apps properly in the target drop down.

It only seems to be necessary to do this for Battleship and HelloWorld. For some reason it's not needed for tutorial-app, despite it seeming to follow the same scheme as Battleship and HelloWorld (perhaps it just worked because it was listed last in the settings file?)

I could change `tutorial/app` to `tutorial/tutorial-app` to keep things consistent, or I could just leave it alone and avoid the rename there (and any ensuing merge conflicts).